Extract code for markdown `code` print
