+++ src/language-markdown/embed.js
@@ -1,6 +1,6 @@
 import { hardline, markAsRoot, replaceEndOfLine } from "../document/index.js";
-import getMaxContinuousCount from "../utilities/get-max-continuous-count.js";
 import inferParser from "../utilities/infer-parser.js";
+import { printCodeFences } from "./print/code.js";
 import { getFencedCodeBlockValue } from "./utilities.js";
 
 function embed(path, options) {
@@ -8,6 +8,10 @@ function embed(path, options) {
 
   switch (node.type) {
     case "code": {
+      if (node.isIndented) {
+        return;
+      }
+
       const { lang: language } = node;
       if (!language) {
         return;
@@ -47,10 +51,7 @@ function embed(path, options) {
           textToDocOptions,
         );
 
-        const styleUnit = options.__inJsTemplate ? "~" : "`";
-        const style = styleUnit.repeat(
-          Math.max(3, getMaxContinuousCount(node.value, styleUnit) + 1),
-        );
+        const style = printCodeFences(path, options);
 
         return markAsRoot([
           style,
