+++ src/language-markdown/print/mdast.js
@@ -13,19 +13,18 @@ import {
   replaceEndOfLine,
   softline,
 } from "../../document/index.js";
-import getMaxContinuousCount from "../../utilities/get-max-continuous-count.js";
 import getMinNotPresentContinuousCount from "../../utilities/get-min-not-present-continuous-count.js";
 import { getPreferredQuote } from "../../utilities/get-preferred-quote.js";
 import UnexpectedNodeError from "../../utilities/unexpected-node-error.js";
 import { locEnd, locStart } from "../loc.js";
 import {
-  getFencedCodeBlockValue,
   getNthListSiblingIndex,
   isAutolink,
   isPrettierIgnore,
   splitText,
 } from "../utilities.js";
 import { printChildren } from "./children.js";
+import { printCode } from "./code.js";
 import { printHeading } from "./heading.js";
 import { printList, printListLegacy } from "./list.js";
 import { printParagraph } from "./paragraph.js";
@@ -213,36 +212,8 @@ function printMdast(path, options, print) {
       return ["> ", align("> ", printChildren(path, options, print))];
     case "heading":
       return printHeading(path, options, print);
-    case "code": {
-      if (node.isIndented) {
-        // indented code block
-        const alignment = " ".repeat(4);
-        return align(alignment, [
-          alignment,
-          replaceEndOfLine(node.value, hardline),
-        ]);
-      }
-
-      // fenced code block
-      const styleUnit = options.__inJsTemplate ? "~" : "`";
-      const style = styleUnit.repeat(
-        Math.max(3, getMaxContinuousCount(node.value, styleUnit) + 1),
-      );
-      return [
-        style,
-        node.lang || "",
-        node.meta ? " " + node.meta : "",
-        hardline,
-        replaceEndOfLine(
-          options.parser === "mdx"
-            ? getFencedCodeBlockValue(node, options.originalText)
-            : node.value,
-          hardline,
-        ),
-        hardline,
-        style,
-      ];
-    }
+    case "code":
+      return printCode(path, options);
     case "html": {
       const { parent, isLast } = path;
       const value =
